\name{BART-package}
\alias{BART-package}
\alias{BART}
\docType{package}
\title{Bayesian Additive Regression Trees}
\description{
Bayesian Additive Regression Trees (BART) provide flexible
nonparametric modeling of covariates for continuous, binary and
time-to-event outcomes.  For more information on BART, see Chipman,
George and McCulloch (2010) <doi:10.1214/09-AOAS285> and Sparapani,
Logan, McCulloch and Laud (2016) <doi:10.1002/sim.6893>.
}
\author{Robert McCulloch: \email{robert.e.mcculloch@gmail.com},\cr
Rodney Sparapani: \email{rsparapa@mcw.edu},\cr
Robert Gramacy: \email{rbg@vt.edu},\cr
Charles Spanbauer: \email{cspanbauer@mcw.edu} and\cr
Matthew Pratola: \email{mpratola@stat.osu.edu}. }
\section{Maintainer}{Rodney Sparapani: \email{rsparapa@mcw.edu}}
\keyword{package}
