\name{BART-package}
\alias{BART-package}
\alias{BART}
\docType{package}
\title{Bayesian Additive Regression Trees}
\description{
To avoid duplication, the main references that this package relies upon appear here only. For more information see Sparapani, Spanbauer and McCulloch <doi:10.18637/jss.v097.i01>. 
}
%% \author{Robert McCulloch: \email{robert.e.mcculloch@gmail.com},\cr
%% Rodney Sparapani: \email{rsparapa@mcw.edu},\cr
%% Robert Gramacy: \email{rbg@vt.edu},\cr
%% Charles Spanbauer: \email{cspanbauer@mcw.edu} and\cr
%% Matthew Pratola: \email{mpratola@stat.osu.edu}. }
%%\section{Maintainer}{Rodney Sparapani: \email{rsparapa@mcw.edu}}
\references{
  Sparapani R., Spanbauer C. and McCulloch R. (2021)
  Nonparametric Machine Learning and Efficient Computation with Bayesian Additive Regression Trees: The BART R Package.
  \emph{JSS}, \bold{97}, 1-66.
  <doi:10.18637/jss.v097.i01>.

  Chipman H., George E. and McCulloch R. (1998)
  Bayesian CART Model Search.
  \emph{JASA}, \bold{93}, 935-948.
  <doi:10.1080/01621459.1998.10473750>.

  Chipman H., George E., and McCulloch R. (2010)
  Bayesian Additive Regression Trees.
  \emph{Annals of Applied Statistics}, \bold{4}, 266-298.
  <doi:10.1214/09-AOAS285>.
  
  Sparapani R., Logan B., McCulloch R. and Laud P. (2016)
  Nonparametric Survival Analysis Using Bayesian Additive Regression Trees (BART).
  \emph{Statistics in Medicine}, \bold{35}, 2741-2753.
  <doi:10.1002/sim.6893>.

  Sparapani R., Logan B., McCulloch R. and Laud P. (2020)
  Nonparametric Competing Risks Analysis Using Bayesian Additive Regression Trees (BART).
  \emph{SMMR}, \bold{29}, 57-77.
  <doi:10.1177/0962280218822140>.
  
  Sparapani R., Rein L., Tarima S., Jackson T. and Meurer J. (2020)
  Non-Parametric Recurrent Events Analysis with BART and an Application to the Hospital Admissions of Patients with Diabetes.
  \emph{Biostatistics}, \bold{21}, 69-85.
  <doi:10.1093/biostatistics/kxy032>.

  Gramacy R. and Polson N. (2012)
  Simulation-based regularized logistic regression.
  \emph{Bayesian Analysis}, \bold{7}, 567-590.
  <doi:10.1214/12-ba719>.
  
  Albert J. and Chib S. (1993)
  Bayesian Analysis of Binary and Polychotomous Response Data.
  \emph{JASA}, \bold{88}, 669-679.
  <doi:10.1080/01621459.1993.10476321>.

  De Waal T., Pannekoek J. and Scholtus S. (2011)
  Handbook of statistical data editing and imputation.
  John Wiley & Sons, Hoboken, NJ.
  
  Friedman J. (1991)
  Multivariate adaptive regression splines.
  \emph{Annals of Statistics}, \bold{19}, 1-67.
  
  Friedman J. (2001)
  Greedy Function Approximation: A Gradient Boosting Machine.
  \emph{Annals of Statistics}, \bold{29}, 1189-1232. 

  Holmes C. and Held L. (2006)
  Bayesian auxiliary variable models for binary and multinomial regression.
  \emph{Bayesian Analysis}, \bold{1}, 145-168.
  <doi:10.1214/06-ba105>.

  Linero A. (2018)
  Bayesian regression trees for high dimensional prediction and variable selection.
  \emph{JASA}, \bold{113}, 626-636.
  <doi:10.1080/01621459.2016.1264957>.
}

\keyword{package}
